#0: sfpu typecast for GS and stub for WHB0   : fix WHB0 issues for typecast float -> uint32, uint16   : to_uint16, to_uint32 types in sfputypes enum
