+++ tt_metal/hw/ckernels/grayskull/metal/llk_api/llk_sfpu/llk_math_eltwise_unary_sfpu_common_includes.h
@@ -101,10 +101,6 @@ inline void llk_math_calculate_sfpu(
         calculate_atan<APPROXIMATION_MODE, ITERATIONS>();
     } else if constexpr (operation == SfpuType::signbit) {
         calculate_signbit<APPROXIMATION_MODE, ITERATIONS>();
-    }else if constexpr (operation == SfpuType::to_uint16) {
-        calculate_to_uint16<APPROXIMATION_MODE, ITERATIONS>();
-    }else if constexpr (operation == SfpuType::to_uint32) {
-        calculate_to_uint32<APPROXIMATION_MODE, ITERATIONS>();
     }else if constexpr (operation == SfpuType::silu) {
         calculate_silu<APPROXIMATION_MODE, ITERATIONS>();
     }
