+++ tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_math_unary_sfpu_api.h
@@ -59,6 +59,26 @@ inline void llk_math_eltwise_unary_sfpu_signbit_init() {
     llk_math_eltwise_unary_sfpu_init<SfpuType::signbit, APPROXIMATE>();
 }
 
+template <bool APPROXIMATE, DstSync dst_sync = DstSync::SyncFull>
+inline void llk_math_eltwise_unary_sfpu_to_uint16(uint dst_index, int vector_mode = VectorMode::RC) {
+    llk_math_eltwise_unary_sfpu<SfpuType::to_uint16, APPROXIMATE, dst_sync>(dst_index, vector_mode);
+}
+
+template <bool APPROXIMATE>
+inline void llk_math_eltwise_unary_sfpu_to_uint16_init() {
+    llk_math_eltwise_unary_sfpu_init<SfpuType::to_uint16, APPROXIMATE>();
+}
+
+template <bool APPROXIMATE, DstSync dst_sync = DstSync::SyncFull>
+inline void llk_math_eltwise_unary_sfpu_to_uint32(uint dst_index, int vector_mode = VectorMode::RC) {
+    llk_math_eltwise_unary_sfpu<SfpuType::to_uint32, APPROXIMATE, dst_sync>(dst_index, vector_mode);
+}
+
+template <bool APPROXIMATE>
+inline void llk_math_eltwise_unary_sfpu_to_uint32_init() {
+    llk_math_eltwise_unary_sfpu_init<SfpuType::to_uint32, APPROXIMATE>();
+}
+
 template <bool APPROXIMATE>
 inline void llk_math_eltwise_unary_sfpu_tanh_init() {
     llk_math_eltwise_unary_sfpu_init<SfpuType::tanh, APPROXIMATE>();
